requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
onnx==1.10.2
pillow==9.3.0
onnxruntime==1.11.1
protobuf==3.20.2
pytest==7.1.2
opencv-python-headless
openvino==2022.2.0
